chore: move non critical functions to a utils file
