+++ packages/element-lib/src/sidetree-v2/protocol/create.js
@@ -1,35 +1,7 @@
-const base64url = require('base64url');
 const MerkleTools = require('merkle-tools');
-const crypto = require('crypto');
+const { encodeJson, decodeJson, getDidUniqueSuffix } = require('../utils');
 
 // TODO deterministic stringify
-const encodeJson = payload => base64url.encode(Buffer.from(JSON.stringify(payload)));
-
-const decodeJson = encodedPayload => JSON.parse(base64url.decode(encodedPayload));
-
-const payloadToHash = (payload) => {
-  const encodedPayload = encodeJson(payload);
-  return base64url.encode(
-    crypto
-      .createHash('sha256')
-      .update(base64url.toBuffer(encodedPayload))
-      .digest(),
-  );
-};
-
-const getDidUniqueSuffix = (decodedOperation) => {
-  switch (decodedOperation.header.operation) {
-    case 'create':
-      return payloadToHash(decodedOperation.decodedPayload);
-    case 'update':
-    case 'recover':
-    case 'delete':
-      return decodedOperation.decodedPayload.didUniqueSuffix;
-    default:
-      throw Error(`Cannot extract didUniqueSuffixe from: ${decodedOperation}`);
-  }
-};
-
 const create = sidetree => async (req) => {
   const requests = Array.isArray(req) ? req : [req];
   const operations = requests.map(encodeJson);
